requirements.txt
torch>=1.12.0
transformers>=4.35.0
Pillow>=9.0.0
pycocotools>=2.0.6
tqdm>=4.64.0
matplotlib>=3.5.0
numpy>=1.21.0
langchain>=0.1.0
langchain-ollama>=0.1.0
scikit-learn>=1.0.0